complete unit tests
