+++ project_one/README.md
@@ -1,3 +1,16 @@
 Run Integration test  
 
-./node_modules/.bin/cypress open
+./node_modules/.bin/cypress open  
+
+What I have a done:  
+-Add CSS styling beyond what is required
+-Write five unit tests
+-Write two integration tests
+-Deploy your project to a live URL
+-Your CSS styling is mobile-first
+-Have a Pull Request in the github repo history (beyond the pull request you use to submit)
+-install & use eslint on your project
+
+
+How to:
+run unit test --> npm test
