+++ project_one/index.html
@@ -69,6 +69,7 @@
       </form>
     </main>
 
+    <script src="js/utils.js" type="text/javascript"></script>
     <script src="js/index.js"></script>
   </body>
 </html>
